+++ src/frame/frame_prepare.rs
@@ -26,7 +26,7 @@ impl Frame {
     /// **Note:** This function should be used internally for building query request frames.
     pub fn new_req_prepare(query: String, flags: Vec<Flag>) -> Frame {
         let version = Version::Request;
-        let stream = rand::random::<u64>();
+        let stream = rand::random::<u16>();
         let opcode = Opcode::Prepare;
         let body = BodyReqPrepare::new(query);
