+++ src/frame/frame_query.rs
@@ -303,7 +303,7 @@ impl Frame {
                          flags: Vec<Flag>)
                          -> Frame {
         let version = Version::Request;
-        let stream = rand::random::<u64>();
+        let stream = rand::random::<u16>();
         let opcode = Opcode::Query;
         let body = BodyReqQuery::new(query,
                                      consistency,
